+++ accounts/managers.py
@@ -13,7 +13,7 @@ class CustomUserManager(BaseUserManager):
             email=self.normalize_email(email).lower(), user_name=user_name, **extra_fields
         )
         user.set_password(password)
-        user.save()
+        user.save(using=self._db)
         return user
 
     def create_superuser(self, email, user_name, password, **extra_fields):